requirements.txt
-i https://pypi.org/simple
alembic==1.4.2
babel==2.8.0
bcrypt==3.1.7
blinker==1.4
cffi==1.14.0
click==7.1.2
dnspython==1.16.0
dominate==2.5.1
email-validator==1.1.1
flask-babelex==0.9.4
flask-bootstrap4==4.0.2
flask-login==0.5.0
flask-mail==0.9.1
flask-migrate==2.5.3
flask-principal==0.4.0
flask-security==3.0.0
flask-sqlalchemy==2.4.3
flask-wtf==0.14.3
flask==1.1.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
jinja2==2.11.2
mako==1.1.3
markupsafe==1.1.1
passlib==1.7.2
psycopg2==2.8.5
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
six==1.15.0
speaklater==1.3
sqlalchemy==1.3.17
visitor==0.1.3
werkzeug==1.0.1
wtforms==2.3.1
